requirements.txt
Django>=4.2,<5.0
djangorestframework
Pillow
django-cors-headers